pyproject.toml
[tool.poetry]
name = "main-project"
version = "0.1.0"
description = ""
authors = ["hwangtate <ghkdxodud10253504@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
django = "^5.1.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
celery = "^5.4.0"
pymysql = "^1.1.1"
django-redis = "^5.4.0"
cryptography = "^43.0.1"
gunicorn = "^23.0.0"
drf-spectacular = "^0.27.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
coverage = "^7.6.1"
mypy = "^1.11.2"
django-stubs = "^5.1.0"
types-pymysql = "^1.1.0.20240524"

[tool.black]
line-length = 140

[tool.isort]
profile = "black"

[tool.coverage.report]
fail_under = 70

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.12
mypy_path = "src"
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.coverage.run]
omit = [
  ".mypy_cache/**/*",
  "*/migrations/*",
  "*/*test*.py",
]

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"